update backport
Made Annex.Exception work without monad-control, of
course it doesn't actually recover from errors without it.
My uses of bracketIO and handle are, so far, only present to handle
exception error conditions and so it's ok to just have the backport of
git-annex throw an exception in these cases.

A few more Applicatives that have to be removed.