fixed Applicative instance for Annex in debian stable
The old mtl doesn't support deriving it, but a manual instance fixes that,
and allows getting rid of all the code rewritten to not be in applicative
style.

The backport diff is back to manageable size, consiting only of dependency
changes, the hack around the lack of monad-control, and the code for
pre-git-1.7 check-attr.